# hash: Add SHA-NI implementation of SHA-256