--- a/README.REDIST.BINS
+++ b/README.REDIST.BINS
@@ -18,7 +18,7 @@
 18. avifinfo (ext/standard/libavifinfo) see ext/standard/libavifinfo/LICENSE
 19. xxHash (ext/hash/xxhash)
 20. Lexbor (ext/dom/lexbor/lexbor) see ext/dom/lexbor/LICENSE
-21. Portions of libcperciva (ext/hash/hash_sha_sse2.c) see the header in the source file
+21. Portions of libcperciva (ext/hash/hash_sha_{ni,sse2}.c) see the header in the source file
 
 3. pcre2lib (ext/pcre)