@cperciva 2024-07-29
@TimWolla I didn't realize you were taking the sse2 version as well as the shani version.  No objection and it's still owned by Tarsnap Backup Inc. but if you're keeping track of authors anywhere the sse2 version was written by @gperciva.