@cperciva 2024-07-29
> I can't test it myself and thus didn't include it, but if we wanted to also add [your ARM implementation](https://github.com/Tarsnap/libcperciva/blob/master/alg/sha256_arm.c) in the future, how would the copyright note need to look like?

Same as the SSE2 version; "Copyright 2021 Tarsnap Backup Inc." with the same BSD license header.  That was also written by Graham.