@bukka 2024-07-29
Wouldn't be better to do the same as Python and rather use OpenSSL implementation (which is already ASM optimized and doesn't require any extra maintenance) and use our own implementation only if OpenSSL is not available (not needing NI)?