@petk 2024-07-29
>  I'm not at all sure if the way I've integrated this with the build process is any good. If someone more experienced with that could give me some helpful pointers, I'd appreciate that. @petk perhaps?

As far as these build system files are concerned no worries, all looks good here. I'll just probably sort them also for the ext/hash extension so the diffs in the future are simpler to handle :D. About the implementation itself, I have no idea if there's anything wrong here. I think it looks good so far. Yes, if this can be added to PHP-8.4 that would be really nice add-on. :+1:  Considering that we've slightly missed the time frame for the BLAKE 3 hash in the 13194 PR by @divinity76.