@TimWolla 2024-07-30
@bukka I don't see how this would cause any extra maintenance after the initial implementation / getting the feature detection to work correctly [1]. I'm intentionally using an existing implementation instead of writing my own, the CPU does the heavy lifting and for a cryptographic hash it's really immediately obvious when the algorithm is incorrectly implemented.

I don't believe that using OpenSSL, which is optional for PHP, would result in any simpler implementation, given that we can't rely on it being there. This PR keeps everything reasonably self-contained.

[1] The latter of which is something we need to understand anyways.