@cmb69 2024-07-30
While I think that @bukka makes good arguments (especially regarding security), there could be a problem with the official Windows builds. ext/hash is statically built in, and so far we refrained from adding dependencies to OpenSSL from the core php8.dll (see https://github.com/php/php-src/pull/5129#issuecomment-581043330). So in practice, Windows users wouldn't benefit from performance improvements. An option might be to add this implementation to winlibs, but then it might get even less attention.