@cmb69 2024-07-31
If we are going to accept this PR, I can finish up the Windows integration (and do some performance tests), but I would like some clarification about https://github.com/php/php-src/pull/15152#discussion_r1695254372 and particularly about https://github.com/php/php-src/pull/15152#discussion_r1695631224. The former is only about the `#ifdef PHP_WIN32` which might be plain wrong (possibly it should be something like `#ifdef __cplusplus`); the latter is likely more complex a subject.