@TimWolla 2024-08-01
> Shouldn't this be included in https://github.com/php/php-src/blob/master/README.REDIST.BINS ?

Thank you, added a reference to that file.