@TimWolla 2024-08-01
Upstream made a small formatting fix after I reported it (https://github.com/Tarsnap/libcperciva/commit/661752aee82dcc8070754e3e539fdc7782bd3942). I've just synced the implementation and used the opportunity to clean up the commit history by squashing the commits and rebasing on the latest master.

From my side this just leaves the Windows integration / the feature detection handling, which cmb mentioned. Unfortunately I can't advice there.