@cmb69 2024-08-07
I've just pushed a commit to replace the remaining uses of `static restrict` with the `PHP_STATIC_RESTRICT` macro.

`static restrict` is definitely not supported by latest MSVC (at least not in this context) even with `/std:c11`. The `#ifdef PHP_WIN32` is likely too strict, since at least in theory it should still possible to build with clang on Windows (although I did not try that for years, and #11313 has been closed without further notice), but we can still fix the condition later on, if necessary.