@TimWolla 2024-08-08
@cmb69 Thank you. My understanding is that with the current state of this PR, the SSE2 implementation would be functional for Windows, but the SHA-NI implementation not?

Is there anything else you'd like to do for now or do you believe that this PR is in a mergeable state and anything else could be adjusted in a follow-up?

I believe it would be nice being able to include it in Beta 1 to give it some exposure.