@cmb69 2024-08-08
On Windows, you will usually get the SSE2 implementation (official builds already rely on SSE2; only custom builds could switch that off).  And it seems to be impossible get the SHA-NI implementation (unless changing the source code) even if you do AVX2 builds (what we're doing in CI).

Some quick performance tests (modified ext/hash/bench.php to only use "sha256") show a slight improvement for SSE2 (~4.2s), but a huge improvement for SHA-NI (~0.7s), compared to the unoptimized implementation (~4.8s). And with the SHA-NI implementation, "sha256" is way faster than "sha1" or even "md5" what appears to be a very good thing.

> I believe it would be nice being able to include it in Beta 1 to give it some exposure.

Definitely! Even as is, I think many users will profit from it, and with the official Windows builds SHA-NI wouldn't be supported anyway (we may consider to distribute AVX2 builds), and we can still catch up on fixing the AVX/AVX2 detection somewhat later.