@TimWolla 2024-08-08
> And with the SHA-NI implementation, "sha256" is way faster than "sha1" or even "md5" what appears to be a very good thing.

Yes, that was one of the goals of this patch: Give users more reasons to use a more secure hashing algorithm.

------

I take your comment as a confirmation that you are happy with the PR / approve of it then?