@TimWolla 2024-08-08
> but if you're keeping track of authors anywhere

I've added both of your names to the NEWS entry: https://github.com/php/php-src/pull/15152/commits/6f4bc0ab7bd5c7af4fc347585260e1c6069a8fff